+++ packages/filterlist/__stories__/index.stories.tsx
@@ -1,9 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
-
 import {
   Browser,
 } from '@vtaits/react-router-dom-fake-browser';
-
 import { AllFeatures } from './AllFeatures';
 
 const meta: Meta = {